+++ packages/remix/test/fixtures-vite/02-interactive-remix-routing-v2/app/data/routes/folderOrganization/index.tsx
@@ -1,0 +1,3 @@
+import folderOrganization from './folderOrganization';
+
+export default folderOrganization;