+++ packages/redwood/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/redwood
 
+## 2.0.8
+
+### Patch Changes
+
+- bump `@vercel/nft@0.26.4` ([#11155](https://github.com/vercel/vercel/pull/11155))
+
+## 2.0.7
+
+### Patch Changes
+
+- [node][next][redwood][remix] bump `@vercel/nft@0.26.3` ([#11115](https://github.com/vercel/vercel/pull/11115))
+
 ## 2.0.6
 
 ### Patch Changes